refactor: gateway api and connectors
