+++ gateway/ui/src/api/Explorer.ts
@@ -1,42 +1,4 @@
-export interface Explorer {
-	// Try get transaction status for a tx with given address
-	watchAddressForTransaction: (
-		address: string
-	) => Promise<TransactionStatus | null>;
-
-	// Try get transaction status for a tx with given address and transaction id.
-	// Since we know the transaction id, this should be more efficient than watchAddressForTransaction
-	watchTransactionStatus: (
-		address: string,
-		txid: string
-	) => Promise<TransactionStatus>;
-
-	// Try get transaction proof for a tx with given transaction id
-	fetchTransactionProof: (txid: string) => Promise<TransactionProof>;
-
-	// Fetch the block height at current chain tip
-	fetchLatestBlockHeight: () => Promise<number>;
-}
-
-export interface TransactionStatus {
-	/** transaction id */
-	transactionId: string;
-	/** amount detected in transaction */
-	amount_btc: number;
-	/** number of confirmations */
-	confirmations: number;
-	/** status of the transaction */
-	status: 'pending' | 'confirmed';
-	/** Url to view this transaction on the explorer where status information was sourced */
-	viewTransactionUrl: string;
-}
-
-export interface TransactionProof {
-	/** Merkle proof of transaction */
-	transactionOutProof: string;
-	/** Sha256 hash of transaction */
-	transactionHash: string;
-}
+import { Explorer, TransactionStatus, TransactionProof } from './api.types';
 
 interface ChaintipCache {
 	// block height at current chain tip
