+++ gateway/ui/src/api/Mintgate.ts
@@ -1,48 +1,100 @@
+import { GatewayInfo, Mintgate } from './api.types';
 import { Federation } from '../federation.types';
-
-// Mintgate is an API to interact with the Gateway server
-export interface Mintgate {
-	fetchInfo: () => Promise<GatewayInfo>;
-	fetchAddress: () => Promise<string>;
-	connectFederation: (connectInfo: string) => Promise<Federation>;
-	/**
-	 * Complete a deposit to a federation served by the Gateway
-	 * @param federationId id of the federation to deposit to
-	 * @param txOutProof transaction out proof for the deposit made to address previously sourced from federation via fetchAddress
-	 * @param tx transaction hash for the deposit made to address previously sourced from federation via fetchAddress
-	 * @returns `TransactionId` from the Fedimint federation
-	 */
-	completeDeposit: (
-		federationId: string,
-		txOutProof: string,
-		tx: string
-	) => Promise<string>;
-
-	/**
-	 *  Request a withdrawal from a fedration served by the Gateway
-	 * @param federationId  id of the federation to withdraw from
-	 * @param amountSat the amount in satoshis to be withdrawn from the federation
-	 * @param address the bitcoin address to withdraw to
-	 * @returns `TransactionId` from the Fedimint federation
-	 */
-	requestWithdrawal: (
-		federationId: string,
-		amountSat: number,
-		address: string
-	) => Promise<string>;
-}
-
-// GatewayInfo is the information returned by the Gateway server
-export interface GatewayInfo {
-	version_hash: string;
-	federations: Federation[];
-}
-
-/** TransactionId of a fedimint federation */
-export type TransactionId = string;
 
 // NullGatewayInfo is a placeholder for when the GatewayInfo is not yet loaded
 export const NullGatewayInfo: GatewayInfo = {
 	version_hash: '',
 	federations: [],
 };
+
+// MockMintgate is a mock implementation of the Mintgate API
+export class MockMintgate implements Mintgate {
+	fetchInfo = async (): Promise<GatewayInfo> => {
+		return {
+			federations: [
+				{
+					federation_id: 'Hals_trusty_mint',
+					mint_pubkey:
+						'4222d0e6f9f4e0a9d471c81b3ee454032cddc8590a9aebf43332bb2d3ecb976c',
+					details: {
+						description: 'Hals Trusty Mint',
+						date_created: '2018-01-01T00:00:00Z',
+						name: 'Hals_trusty_mint',
+						url: '',
+						active: true,
+					},
+				},
+
+				{
+					federation_id: 'FTX_yield_gen',
+					mint_pubkey:
+						'c679bce3d2bb23334fbea9a0958cddc230454ee3b18c174d9a0e4f9f6e0d2224',
+					details: {
+						description: 'FTX yield generator',
+						date_created: '2019-01-01T00:00:00Z',
+						name: 'FTX_yield_gen',
+						url: '',
+						active: true,
+					},
+				},
+
+				{
+					federation_id: 'Qala_devs',
+					mint_pubkey:
+						'1d2eb31bb4c307b466e1a7c910693bc5754799ec705b14f760e74d5d4968acf9',
+					details: {
+						description: 'Qala_devs stacking sats',
+						date_created: '2020-01-01T00:00:00Z',
+						name: 'Qala_devs',
+						url: '',
+						active: false,
+					},
+				},
+			],
+			version_hash: 'f876d12a78df7d4ef5b3a4913590ded6bfe5c3b1',
+		};
+	};
+
+	fetchAddress = async (): Promise<string> => {
+		// Use IP's bitcoin address as a mock
+		return 'bc1qgf60crqtlxn7279tgh8lsxzagmu97cyuwtykxwv026s9hwg427fsjvw7uz';
+	};
+
+	connectFederation = async (_connectInfo: string): Promise<Federation> => {
+		// TODO: autogenerate mock names
+		const federation_id = 'mock_federation';
+		// TODO: autogenerate mock pubkeys
+		const mint_pubkey =
+			'4222d0e6f9f4e0a9d471c81b3ee454032cddc8590a9aebf43332bb2d3ecb976c';
+
+		return Promise.resolve({
+			federation_id,
+			mint_pubkey,
+			details: {
+				description: 'Newly connected federation mock',
+				date_created: Date.now().toString(),
+				name: 'mock_federation',
+				url: '',
+				active: true,
+			},
+		});
+	};
+
+	completeDeposit = async (
+		_federationId: string,
+		_txOutProof: string,
+		_tx: string
+	): Promise<string> => {
+		return Promise.resolve('mock_txid');
+	};
+
+	requestWithdrawal = async (
+		_federationId: string,
+		_amountSat: number,
+		_address: string
+	): Promise<string> => {
+		return Promise.resolve(
+			'de3d5bf1e3c1b3be2a1e025825f751629390ad60c8f91723e330f2356d99c59b'
+		);
+	};
+}
